% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_kon_alder.R
\name{ps_e_kon_alder}
\alias{ps_e_kon_alder}
\title{Partisympati efter kön och ålder, 4 åldersklasser (urvalsundersökning).}
\usage{
ps_e_kon_alder()
}
\description{
Partisympati efter kön och ålder, 4 åldersklasser (urvalsundersökning).
}
