% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_kon_alder_7klasser.R
\name{ps_e_kon_alder_7klasser}
\alias{ps_e_kon_alder_7klasser}
\title{Partisympati efter kön och ålder, 7 åldersklasser (urvalsundersökning).}
\usage{
ps_e_kon_alder_7klasser()
}
\description{
Partisympati efter kön och ålder, 7 åldersklasser (urvalsundersökning).
}
