% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_kon_bostadstyp.R
\name{ps_e_kon_bostadstyp}
\alias{ps_e_kon_bostadstyp}
\title{Partisympati efter kön och bostadstyp (urvalsundersökning).}
\usage{
ps_e_kon_bostadstyp()
}
\description{
Partisympati efter kön och bostadstyp (urvalsundersökning).
}
