% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_kon_civilstand.R
\name{ps_e_kon_civilstand}
\alias{ps_e_kon_civilstand}
\title{Partisympati efter kön och civilstånd (urvalsundersökning).}
\usage{
ps_e_kon_civilstand()
}
\description{
Partisympati efter kön och civilstånd (urvalsundersökning).
}
