% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_kon_fackforbund.R
\name{ps_e_kon_fackforbund}
\alias{ps_e_kon_fackforbund}
\title{Partisympati efter kön och fackförbund bland anställda (urvalsundersökning).}
\usage{
ps_e_kon_fackforbund()
}
\description{
Partisympati efter kön och fackförbund bland anställda (urvalsundersökning).
}
