% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_kon_inkomst.R
\name{ps_e_kon_inkomst}
\alias{ps_e_kon_inkomst}
\title{Partisympati efter kön och inkomstintervall efter percentiler (urvalsundersökning).}
\usage{
ps_e_kon_inkomst()
}
\description{
Partisympati efter kön och inkomstintervall efter percentiler (urvalsundersökning).
}
