% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_kon_med_utan_barn.R
\name{ps_e_kon_med_utan_barn}
\alias{ps_e_kon_med_utan_barn}
\title{Partisympati efter kön och utan/med barn (urvalsundersökning).}
\usage{
ps_e_kon_med_utan_barn()
}
\description{
Partisympati efter kön och utan/med barn (urvalsundersökning).
}
