% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_kon_region.R
\name{ps_e_kon_region}
\alias{ps_e_kon_region}
\title{Partisympati efter kön och region (Sverige, 8 grupper) (urvalsundersökning).}
\usage{
ps_e_kon_region()
}
\description{
Partisympati efter kön och region (Sverige, 8 grupper) (urvalsundersökning).
}
