% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_kon_sektor_anstallda.R
\name{ps_e_kon_sektor_anstallda}
\alias{ps_e_kon_sektor_anstallda}
\title{Partisympati efter kön och sektor bland sysselsatta (urvalsundersökning).}
\usage{
ps_e_kon_sektor_anstallda()
}
\description{
Partisympati efter kön och sektor bland sysselsatta (urvalsundersökning).
}
