% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_kon_socioekonomiskIndelning.R
\name{ps_e_kon_socioekonomiskIndelning}
\alias{ps_e_kon_socioekonomiskIndelning}
\title{Partisympati efter kön och socioekonomisk indelning (SEI) i väljarkåren (urvalsundersökning).}
\usage{
ps_e_kon_socioekonomiskIndelning()
}
\description{
Partisympati efter kön och socioekonomisk indelning (SEI) i väljarkåren (urvalsundersökning).
}
