% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_kon_sysselsatt.R
\name{ps_e_kon_sysselsatt}
\alias{ps_e_kon_sysselsatt}
\title{Partisympati efter kön och sysselsatt (urvalsundersökning).}
\usage{
ps_e_kon_sysselsatt()
}
\description{
Partisympati efter kön och sysselsatt (urvalsundersökning).
}
