% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_kon_utbildning.R
\name{ps_e_kon_utbildning}
\alias{ps_e_kon_utbildning}
\title{Partisympati efter kön och utbildningsnivå (urvalsundersökning).}
\usage{
ps_e_kon_utbildning()
}
\description{
Partisympati efter kön och utbildningsnivå (urvalsundersökning).
}
