% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_kon_utrikes_inrikes_bakgrund.R
\name{ps_e_kon_utrikes_inrikes_bakgrund}
\alias{ps_e_kon_utrikes_inrikes_bakgrund}
\title{Partisympati efter kön och utländsk/svensk bakgrund (urvalsundersökning).}
\usage{
ps_e_kon_utrikes_inrikes_bakgrund()
}
\description{
Partisympati efter kön och utländsk/svensk bakgrund (urvalsundersökning).
}
