% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_kon_utrikes_inrikes_fodd.R
\name{ps_e_kon_utrikes_inrikes_fodd}
\alias{ps_e_kon_utrikes_inrikes_fodd}
\title{Partisympati efter kön och utrikes/inrikes född (urvalsundersökning).}
\usage{
ps_e_kon_utrikes_inrikes_fodd()
}
\description{
Partisympati efter kön och utrikes/inrikes född (urvalsundersökning).
}
