% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_kon_yrkesklassifisering.R
\name{ps_e_kon_yrkesklassifisering}
\alias{ps_e_kon_yrkesklassifisering}
\title{Partisympati efter kön och standard för svensk yrkesklassificering (SSYK 2012) i väljarkåren.}
\usage{
ps_e_kon_yrkesklassifisering()
}
\description{
Partisympati efter kön och standard för svensk yrkesklassificering (SSYK 2012) i väljarkåren.
}
