% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_nast_basta_parti.R
\name{ps_e_nast_basta_parti}
\alias{ps_e_nast_basta_parti}
\title{Väljarkåren (svarsfördelning PSU) efter partisympati, näst bästa parti}
\usage{
ps_e_nast_basta_parti()
}
\description{
Väljarkåren (svarsfördelning PSU) efter partisympati, näst bästa parti
}
