% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_region.R
\name{ps_e_region}
\alias{ps_e_region}
\title{Partisympati efter region (Sverige, 10 grupper) (urvalsundersökning).}
\usage{
ps_e_region()
}
\description{
Partisympati efter region (Sverige, 10 grupper) (urvalsundersökning).
}
