% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_sektor_anstallda.R
\name{ps_e_sektor_anstallda}
\alias{ps_e_sektor_anstallda}
\title{Partisympati efter sektor bland sysselsatta (urvalsundersökning). 2 ggr/år 2006M05 - 2021M05}
\usage{
ps_e_sektor_anstallda()
}
\description{
Partisympati efter sektor bland sysselsatta (urvalsundersökning). 2 ggr/år 2006M05 - 2021M05
}
