% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_e_yrkesklassifisering.R
\name{ps_e_yrkesklassifisering}
\alias{ps_e_yrkesklassifisering}
\title{Partisympati efter standard för svensk yrkesklassificering (SSYK 2012) bland sysselsatta.}
\usage{
ps_e_yrkesklassifisering()
}
\description{
Partisympati efter standard för svensk yrkesklassificering (SSYK 2012) bland sysselsatta.
}
