% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sympati_basta_parti.R
\name{sympati_basta_parti}
\alias{sympati_basta_parti}
\title{Röstningssympati (parti man skulle rösta på) efter partisympati (bästa parti).}
\usage{
sympati_basta_parti(tid)
}
\description{
Röstningssympati (parti man skulle rösta på) efter partisympati (bästa parti).
}
