% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sympati_parti_att_rosta_pa.R
\name{sympati_parti_att_rosta_pa}
\alias{sympati_parti_att_rosta_pa}
\title{Röstningssympati (parti man skulle rösta på) efter partisympati (bästa parti).}
\usage{
sympati_parti_att_rosta_pa(tid)
}
\description{
Röstningssympati (parti man skulle rösta på) efter partisympati (bästa parti).
}
