% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sympati_parti_efter_forra_PSU.R
\name{sympati_parti_efter_forra_PSU}
\alias{sympati_parti_efter_forra_PSU}
\title{Röstningssympati nu (parti man skulle rösta på) efter röstningssympati i förra PSU.}
\usage{
sympati_parti_efter_forra_PSU(tid)
}
\description{
Röstningssympati nu (parti man skulle rösta på) efter röstningssympati i förra PSU.
}
