% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sympati_parti_efter_rixval.R
\name{sympati_parti_efter_rixval}
\alias{sympati_parti_efter_rixval}
\title{Röstningssympati nu (parti man skulle rösta på) efter partival i förra riksdagsvalet.}
\usage{
sympati_parti_efter_rixval(tid)
}
\description{
Röstningssympati nu (parti man skulle rösta på) efter partival i förra riksdagsvalet.
}
