% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valresultat_val_idag_kon.R
\name{valresultat_val_idag_kon}
\alias{valresultat_val_idag_kon}
\title{Valresultat om det varit val idag (PSU) efter kön.}
\usage{
valresultat_val_idag_kon(tid)
}
\description{
Valresultat om det varit val idag (PSU) efter kön.
}
