% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valresultat_val_idag_region.R
\name{valresultat_val_idag_region}
\alias{valresultat_val_idag_region}
\title{Valresultat om det varit val idag (PSU) efter region (Sverige, 8 grupper).}
\usage{
valresultat_val_idag_region(tid)
}
\description{
Valresultat om det varit val idag (PSU) efter region (Sverige, 8 grupper).
}
