% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valresutlat_val_idag.R
\name{valresutlat_val_idag}
\alias{valresutlat_val_idag}
\title{Valresultat om det varit val idag (PSU).}
\usage{
valresutlat_val_idag(tid)
}
\description{
Valresultat om det varit val idag (PSU).
}
